+++ src/components/UI/input/FormInput/index.tsx
@@ -1,35 +1,35 @@
-import React, { FC, RefObject, RefAttributes, ChangeEvent } from 'react';
+import React, { FC } from 'react';
+import { FormValues } from 'components/FormPerson/FormPerson.interface';
+import { UseFormRegister, FieldError } from 'react-hook-form';
 import cl from './FormInput.module.scss';
 
-interface InputProps {
-  type: string;
-  placeholder: string;
+type InputProps = {
   label: string;
-  name: string;
-  error: string | undefined;
-  onChange: (event: ChangeEvent<HTMLInputElement>) => void;
-  ref: RefObject<HTMLInputElement>;
-}
+  type: string;
+  placeholder?: string;
+  register: ReturnType<UseFormRegister<FormValues>>;
+  error: FieldError | undefined;
+};
 
-const FormInput: FC<Omit<InputProps, 'ref'> & RefAttributes<HTMLInputElement>> = React.forwardRef(
-  ({ placeholder, type, label, name, error, onChange }, ref) => {
-    return (
-      <>
-        <label className={cl.labelInput}>
-          {label}
-          <input
-            ref={ref}
-            name={name}
-            type={type}
-            placeholder={placeholder}
-            className={cl.formInput}
-            onChange={onChange}
-          />
-          {error ? <div className={cl.error}>{error}</div> : <div className={cl.noerror}></div>}
-        </label>
-      </>
-    );
-  }
-);
+const FormHookInput: FC<InputProps> = ({ label, type, placeholder, register, error }) => {
+  return (
+    <>
+      <label className={type === 'checkbox' ? cl.labelCheck : cl.labelInput}>
+        {label}
+        <input
+          {...register}
+          type={type}
+          placeholder={placeholder}
+          className={type === 'checkbox' ? cl.inputCheck : cl.formInput}
+        />
+        {error ? (
+          <div className={cl.error}>{error.message}</div>
+        ) : (
+          <div className={cl.noerror}></div>
+        )}
+      </label>
+    </>
+  );
+};
 
-export default FormInput;
+export default FormHookInput;
