+++ src/pages/FormPage/index.tsx
@@ -1,37 +1,25 @@
-import React, { Component } from 'react';
+import React, { FC, useState } from 'react';
 import { FormValues } from 'components/FormPerson/FormPerson.interface';
 import './FormPage.scss';
 import PersonCards from 'components/PersonCards';
-import FormHookPerson from 'components/FormPerson/FormHookPerson';
+import FormHookPerson from 'components/FormPerson';
 
-interface FormPageState {
-  personCards: FormValues[] | [];
-}
+const FormHookPage: FC = () => {
+  const [personCards, setPersonCards] = useState<FormValues[]>([]);
 
-export class index extends Component<Record<string, unknown>, FormPageState> {
-  constructor(props: Record<string, unknown>) {
-    super(props);
-    this.state = {
-      personCards: [],
-    };
-    this.addCard = this.addCard.bind(this);
-  }
-
-  addCard(card: FormValues) {
-    this.setState((state) => {
-      return { ...state, personCards: [...state.personCards, card] };
+  const addCard = (card: FormValues) => {
+    setPersonCards((prev) => {
+      return [...prev, card];
     });
-  }
-  render() {
-    return (
-      <div className="main__container">
-        <div className="main__content">
-          <FormHookPerson addCard={this.addCard} />
-          <PersonCards personCards={this.state.personCards} />
-        </div>
+  };
+  return (
+    <div className="main__container">
+      <div className="main__content">
+        <FormHookPerson addCard={addCard} />
+        <PersonCards personCards={personCards} />
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
-export default index;
+export default FormHookPage;
