+++ src/pages/Home/Home.test.tsx
@@ -1,72 +1,16 @@
-import { render, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 import React from 'react';
-import { Home } from '.';
+import { render, screen, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import Index from './index';
 
-test('renders filtered products list when search text is entered', async () => {
-  const products = [
-    {
-      id: 1,
-      title: 'iPhone 9',
-      description: 'An apple mobile which is nothing like apple',
-      price: 549,
-      discountPercentage: 12.96,
-      rating: 4.69,
-      stock: 94,
-      brand: 'Apple',
-      category: 'smartphones',
-      thumbnail: 'https://i.dummyjson.com/data/products/1/thumbnail.jpg',
-      images: [
-        'https://i.dummyjson.com/data/products/1/1.jpg',
-        'https://i.dummyjson.com/data/products/1/2.jpg',
-        'https://i.dummyjson.com/data/products/1/3.jpg',
-        'https://i.dummyjson.com/data/products/1/4.jpg',
-        'https://i.dummyjson.com/data/products/1/thumbnail.jpg',
-      ],
-    },
-    {
-      id: 2,
-      title: 'iPhone X',
-      description:
-        'SIM-Free, Model A19211 6.5-inch Super Retina HD display with OLED technology A12 Bionic chip with ...',
-      price: 899,
-      discountPercentage: 17.94,
-      rating: 4.44,
-      stock: 34,
-      brand: 'Apple',
-      category: 'smartphones',
-      thumbnail: 'https://i.dummyjson.com/data/products/2/thumbnail.jpg',
-      images: [
-        'https://i.dummyjson.com/data/products/2/1.jpg',
-        'https://i.dummyjson.com/data/products/2/2.jpg',
-        'https://i.dummyjson.com/data/products/2/3.jpg',
-        'https://i.dummyjson.com/data/products/2/thumbnail.jpg',
-      ],
-    },
-    {
-      id: 3,
-      title: 'Samsung Universe 9',
-      description: "Samsung's new variant which goes beyond Galaxy to the Universe",
-      price: 1249,
-      discountPercentage: 15.46,
-      rating: 4.09,
-      stock: 36,
-      brand: 'Samsung',
-      category: 'smartphones',
-      thumbnail: 'https://i.dummyjson.com/data/products/3/thumbnail.jpg',
-      images: ['https://i.dummyjson.com/data/products/3/1.jpg'],
-    },
-  ];
-  const { getByPlaceholderText, getByText } = render(<Home products={products} />);
-
-  await waitFor(() => {
-    const product1 = getByText(/mobile which/i);
-    const product2 = getByText(/Super Retina/i);
-    const product3 = getByText(/variant which/i);
-    const searchInput = getByPlaceholderText(/Search product/i);
-    userEvent.type(searchInput, 'Super Retina');
-    expect(product1).not.toBeInTheDocument();
-    expect(product2).toBeInTheDocument();
-    expect(product3).not.toBeInTheDocument();
+describe('HomeHook', () => {
+  it('search bar filters products correctly', async () => {
+    render(<Index />);
+    const searchInput = screen.getByPlaceholderText(/Search product/i);
+    await waitFor(() => {
+      userEvent.type(searchInput, 'sams');
+      const productsList = screen.getByTestId('products-list');
+      expect(productsList.children.length).toBe(2);
+    });
   });
 });
