+++ src/pages/Home/index.tsx
@@ -1,54 +1,26 @@
 import ProductsList from 'components/ProductsList';
-import React, { Component } from 'react';
+import React, { FC, useState } from 'react';
 import './Home.scss';
 import products from 'db/products';
 import { Product } from '../../models/product';
-// import SearchBar from 'components/SearchBar';
-import SearchHookBar from 'components/SearchBar/SearchHookBar';
+import SearchHookBar from 'components/SearchBar';
 
-interface HomeState {
-  products: Product[];
-  searchedProducts: Product[];
-}
-interface HomeProps {
-  products?: Product[];
-}
-
-export class Home extends Component<HomeProps, HomeState> {
-  constructor(props: HomeProps) {
-    super(props);
-    this.state = {
-      products: products,
-      searchedProducts: products,
-    };
-    this.setSearchedProducts = this.setSearchedProducts.bind(this);
-  }
-
-  setSearchedProducts(searchedProducts: Product[]) {
-    this.setState((state) => {
-      return { ...state, searchedProducts: searchedProducts };
-    });
-  }
-
-  render() {
-    return (
-      <section className="main__section">
-        <div className="main__container">
-          <div className="main__content">
-            <SearchHookBar
-              products={this.state.products}
-              setSearchedProducts={this.setSearchedProducts}
-            />
-            {this.state.searchedProducts.length ? (
-              <ProductsList products={this.state.searchedProducts} />
-            ) : (
-              <h2> Товары не найдены</h2>
-            )}
-          </div>
+const HomeHook: FC = () => {
+  const [searchedProducts, setSearchedProducts] = useState<Product[]>([]);
+  return (
+    <section className="main__section">
+      <div className="main__container">
+        <div className="main__content">
+          <SearchHookBar products={products} setSearchedProducts={setSearchedProducts} />
+          {searchedProducts.length ? (
+            <ProductsList products={searchedProducts} />
+          ) : (
+            <h2> Товары не найдены</h2>
+          )}
         </div>
-      </section>
-    );
-  }
-}
+      </div>
+    </section>
+  );
+};
 
-export default Home;
+export default HomeHook;
